sql: expose module structure
Introduce a `plan` module to the `sql` crate, then expose this structure
publically. This gives module documentation a nice place to live.
