+++ src/coord/src/util.rs
@@ -7,7 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use sql::Session;
+use sql::session::Session;
 
 use crate::command::Response;
