+++ src/sql/src/lib.rs
@@ -11,227 +11,16 @@
 
 #![deny(missing_debug_implementations)]
 
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
-
-use ::expr::{GlobalId, RowSetFinishing};
-use dataflow_types::{PeekWhen, SinkConnectorBuilder, SourceConnector, Timestamp};
-use repr::{RelationDesc, Row, ScalarType};
-use sql_parser::parser;
-
-use crate::catalog::Catalog;
-
-pub use crate::expr::RelationExpr;
-pub use names::{DatabaseSpecifier, FullName, PartialName};
-pub use pure::purify_statement as purify;
-pub use session::{PreparedStatement, Session, TransactionStatus};
-pub use sql_parser::ast::{ExplainOptions, ExplainStage, ObjectType, Statement};
-pub use statement::StatementContext;
-
-pub mod catalog;
-pub mod normalize;
-
-mod explain;
-mod expr;
-mod func;
 mod kafka_util;
-mod names;
-mod pure;
-mod query;
-mod scope;
-mod session;
-mod statement;
-mod transform;
 
-// this is used by sqllogictest to turn sql values into `Datum`
-pub use query::scalar_type_from_sql;
-
-/// Instructions for executing a SQL query.
-#[derive(Debug)]
-pub enum Plan {
-    CreateDatabase {
-        name: String,
-        if_not_exists: bool,
-    },
-    CreateSchema {
-        database_name: DatabaseSpecifier,
-        schema_name: String,
-        if_not_exists: bool,
-    },
-    CreateSource {
-        name: FullName,
-        source: Source,
-        if_not_exists: bool,
-        materialized: bool,
-    },
-    CreateSink {
-        name: FullName,
-        sink: Sink,
-        if_not_exists: bool,
-    },
-    CreateTable {
-        name: FullName,
-        desc: RelationDesc,
-        if_not_exists: bool,
-    },
-    CreateView {
-        name: FullName,
-        view: View,
-        /// The ID of the object that this view is replacing, if any.
-        replace: Option<GlobalId>,
-        /// whether we should auto-materialize the view
-        materialize: bool,
-        if_not_exists: bool,
-    },
-    CreateIndex {
-        name: FullName,
-        index: Index,
-        if_not_exists: bool,
-    },
-    DropDatabase {
-        name: String,
-    },
-    DropSchema {
-        database_name: DatabaseSpecifier,
-        schema_name: String,
-    },
-    DropItems {
-        items: Vec<GlobalId>,
-        ty: ObjectType,
-    },
-    EmptyQuery,
-    ShowAllVariables,
-    ShowVariable(String),
-    SetVariable {
-        name: String,
-        value: String,
-    },
-    StartTransaction,
-    CommitTransaction,
-    AbortTransaction,
-    Peek {
-        source: ::expr::RelationExpr,
-        when: PeekWhen,
-        finishing: RowSetFinishing,
-        materialize: bool,
-    },
-    Tail {
-        id: GlobalId,
-        with_snapshot: bool,
-        ts: Option<Timestamp>,
-    },
-    SendRows(Vec<Row>),
-    ExplainPlan {
-        sql: String,
-        raw_plan: crate::expr::RelationExpr,
-        decorrelated_plan: ::expr::RelationExpr,
-        row_set_finishing: Option<RowSetFinishing>,
-        stage: ExplainStage,
-        options: ExplainOptions,
-    },
-    SendDiffs {
-        id: GlobalId,
-        updates: Vec<(Row, isize)>,
-        affected_rows: usize,
-        kind: MutationKind,
-    },
-    ShowViews {
-        ids: Vec<(String, GlobalId)>,
-        full: bool,
-        show_queryable: bool,
-        limit_materialized: bool,
-    },
-}
-
-#[derive(Clone, Debug)]
-pub struct Source {
-    pub create_sql: String,
-    pub connector: SourceConnector,
-    pub desc: RelationDesc,
-}
-
-#[derive(Clone, Debug)]
-pub struct Sink {
-    pub create_sql: String,
-    pub from: GlobalId,
-    pub connector_builder: SinkConnectorBuilder,
-}
-
-#[derive(Clone, Debug)]
-pub struct View {
-    pub create_sql: String,
-    pub expr: ::expr::RelationExpr,
-    pub desc: RelationDesc,
-    pub temporary: bool,
-}
-
-#[derive(Clone, Debug)]
-pub struct Index {
-    pub create_sql: String,
-    pub on: GlobalId,
-    pub keys: Vec<::expr::ScalarExpr>,
-}
-
-#[derive(Debug)]
-pub enum MutationKind {
-    Insert,
-    Update,
-    Delete,
-}
-
-/// A vector of values to which parameter references should be bound.
-#[derive(Debug, Clone)]
-pub struct Params {
-    pub datums: Row,
-    pub types: Vec<ScalarType>,
-}
-
-/// Controls planning of a SQL query.
-#[derive(Ord, PartialOrd, Clone, Debug, Eq, PartialEq, Serialize, Deserialize, Hash)]
-pub struct PlanContext {
-    pub wall_time: DateTime<Utc>,
-}
-
-impl Default for PlanContext {
-    fn default() -> PlanContext {
-        PlanContext {
-            wall_time: Utc::now(),
-        }
-    }
-}
-
-/// Parses a raw SQL string into a [`Statement`].
-pub fn parse(sql: String) -> Result<Vec<Statement>, failure::Error> {
-    Ok(parser::parse_statements(sql)?)
-}
-
-/// Produces a [`Plan`] from the purified statement `stmt`.
-///
-/// Planning is a pure, synchronous function and so requires that the provided
-/// `stmt` does does not depend on any external state. To purify a statement,
-/// use [`purify`].
-pub fn plan(
-    pcx: &PlanContext,
-    catalog: &dyn Catalog,
-    stmt: Statement,
-    params: &Params,
-) -> Result<Plan, failure::Error> {
-    statement::handle_statement(pcx, catalog, stmt, params)
-}
-
-/// Determines the type of the rows that will be returned by `stmt` and the type
-/// of the parameters required by `stmt`. If the statement will not produce a
-/// result set (e.g., most `CREATE` or `DROP` statements), no `RelationDesc`
-/// will be returned. If the query uses no parameters, then the returned vector
-/// of types will be empty.
-pub fn describe(
-    catalog: &dyn Catalog,
-    stmt: Statement,
-) -> Result<(Option<RelationDesc>, Vec<pgrepr::Type>), failure::Error> {
-    let (desc, types) = statement::describe_statement(catalog, stmt)?;
-    let types = types.into_iter().map(|t| pgrepr::Type::from(&t)).collect();
-    Ok((desc, types))
-}
+pub use sql_parser::ast;
+pub mod catalog;
+pub mod names;
+pub mod normalize;
+pub mod parse;
+pub mod plan;
+pub mod pure;
+pub mod session;
 
 #[macro_export]
 macro_rules! unsupported {
