+++ src/sql/src/plan/scope.rs
