+++ src/sql/tests/parameters.rs
@@ -80,8 +80,8 @@ fn test_parameter_type_inference() -> Result<(), Box<dyn Error>> {
     ];
     for (sql, types) in test_cases {
         println!("> {}", sql);
-        let stmt = sql::parse(sql.into())?.into_element();
-        let (_desc, param_types) = sql::describe(&DummyCatalog, stmt)?;
+        let stmt = sql::parse::parse(sql.into())?.into_element();
+        let (_desc, param_types) = sql::plan::describe(&DummyCatalog, stmt)?;
         assert_eq!(param_types, types);
     }
     Ok(())
